requirements.txt
Pillow
requests
Jinja2
